requirements.txt
cairosvg
chess
pixcat
pyyaml
